pyproject.toml
[tool.poetry]
name = "re-yaml-common"
version = "1.0.0"
description = ""
authors = ["Philipp Gayret <philipp@release-engineers.com>"]
readme = "README.md"
packages = [{ include = "yaml_common" }]
license = "SSPL-1.0"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.scripts]
"yaml-common" = "yaml_common.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
fire = "^0.5.0"
"ruamel.yaml" = "^0.17.21"
deepdiff = "^5.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
